+++ src/dal/remote/factory.py
@@ -2,7 +2,7 @@
 
 
 from src.dal.remote.reddit_adapter import RedditAdapter
-
+from src.core.logs import error
 
 class AdapterFactory:
 
@@ -11,8 +11,9 @@ class AdapterFactory:
     }
 
     @classmethod
-    def get_adapter(cls, source_name: str):
-        adapter_class = cls.adapters.get(source_name.lower())
+    def get_adapter(cls, item_name: str):
+        adapter_class = cls.adapters.get(item_name.lower())
         if not adapter_class:
-            raise ValueError(f"No adapter found for source: {source_name}")
+            error(f"No adapter found for source: {item_name}")
+            return None
         return adapter_class()